Eliminate 'fast inner products' for face and edge properties. Done naturally.
