+++ discretize/base/base_tensor_mesh.py
@@ -990,337 +990,6 @@ class BaseTensorMesh(BaseRegularMesh):
         else:
             return None
 
-    def _fastInnerProductSurface(
-        self, projection_type, model=None, invert_model=False, invert_matrix=False
-    ):
-        """Fast version of get_face_inner_product_surface.
-
-        This does not handle the case of a full tensor property.
-
-        Parameters
-        ----------
-        projection_type : str
-            'edges' or 'faces'
-        model : numpy.ndarray
-            material property (tensor properties are possible) at each cell center (nC, (1, 3, or 6))
-        invert_model : bool
-            inverts the material property
-        invert_matrix : bool
-            inverts the matrix
-
-        Returns
-        -------
-        (n_faces, n_faces) or (n_edges, n_edges) scipy.sparse.csr_matrix
-            M, the inner product matrix
-
-        """
-        projection_type = projection_type[0].upper()
-        if projection_type not in ["F", "E"]:
-            raise ValueError("projection_type must be 'F' for faces or 'E' for edges")
-
-        if model is None:
-            model = np.ones(self.nF)
-
-        if invert_model:
-            model = 1.0 / model
-
-        if is_scalar(model):
-            model = model * np.ones(self.nF)
-        # COULD ADD THIS CASE IF DESIRED
-        # elif len(model) == self.dim:
-        #     model = np.r_[
-        #         [model[ii]*self.vnF[ii] for ii in range(0, self.dim)]
-        #     ]
-
-        # number of elements we are averaging (equals dim for regular
-        # meshes, but for cyl, where we use symmetry, it is 1 for edge
-        # variables and 2 for face variables)
-        if self._meshType == "CYL":
-            shape = getattr(self, "vn" + projection_type)
-            if self.is_symmetric:
-                n_elements = 1
-            else:
-                n_elements = sum([1 if x != 0 else 0 for x in shape]) - 1
-        else:
-            n_elements = self.dim - 1
-
-        # Isotropic case only
-        if model.size == self.nF:
-            Aprop = self.face_areas * mkvc(model)
-            if projection_type == "E":
-                Av = self.average_edge_to_face
-                M = n_elements * sdiag(Av.T * Aprop)
-            else:
-                M = sdiag(Aprop)
-
-        else:
-            raise Exception(
-                "Unexpected shape of tensor: {}".format(model.shape),
-                "Must be scalar or have length equal to total number of faces.",
-            )
-
-        if invert_matrix:
-            return sdinv(M)
-        else:
-            return M
-
-    def _fastInnerProductSurfaceDeriv(
-        self, projection_type, model, invert_model=False, invert_matrix=False
-    ):
-        """Faster function for inner product derivatives on tensor meshes.
-
-        Parameters
-        ----------
-        projection_type : str
-            'edges' or 'faces'
-        model : numpy.ndarray
-            material property (tensor properties are possible) at each cell center (nC, (1, 3, or 6))
-        invert_model : bool
-            inverts the material property
-        invert_matrix : bool
-            inverts the matrix
-
-        Returns
-        -------
-        function
-            dMdmu, the derivative of the inner product matrix
-        """
-        projection_type = projection_type[0].upper()
-        if projection_type not in ["F", "E"]:
-            raise ValueError("projection_type must be 'F' for faces or 'E' for edges")
-
-        if is_scalar(model):
-            tensorType = 0
-        elif model.size == self.nF:
-            tensorType = 1
-        else:
-            raise Exception(
-                "Unexpected shape of tensor: {}".format(model.shape),
-                "Must be scalar or have length equal to total number of faces."
-            )
-
-        dMdprop = None
-
-        if invert_matrix or invert_model:
-            MI = self._fastFacePropertiesInnerProduct(
-                projection_type,
-                model,
-                invert_model=invert_model,
-                invert_matrix=invert_matrix,
-            )
-
-        # number of elements we are averaging (equals dim for regular
-        # meshes, but for cyl, where we use symmetry, it is 1 for edge
-        # variables and 2 for face variables)
-        if self._meshType == "CYL":
-            shape = getattr(self, "vn" + projection_type)
-            if self.is_symmetric:
-                n_elements = 1
-            else:
-                n_elements = sum([1 if x != 0 else 0 for x in shape]) - 1
-        else:
-            n_elements = self.dim - 1
-
-        A = sdiag(self.face_areas)
-        if projection_type == "E":
-            Av = self.average_edge_to_face
-        else:
-            Av = sdiag(np.ones(self.nF) / n_elements)
-
-        if tensorType == 0:  # isotropic, constant
-            ones = sp.csr_matrix(
-                (np.ones(self.nF), (range(self.nF), np.zeros(self.nF))),
-                shape=(self.nF, 1),
-            )
-            if not invert_matrix and not invert_model:
-                dMdprop = n_elements * Av.T * A * ones
-            elif invert_matrix and invert_model:
-                dMdprop = n_elements * (
-                    sdiag(MI.diagonal() ** 2)
-                    * Av.T
-                    * A
-                    * ones
-                    * sdiag(1.0 / model**2)
-                )
-            elif invert_model:
-                dMdprop = n_elements * Av.T * A * sdiag(-1.0 / model**2)
-            elif invert_matrix:
-                dMdprop = n_elements * (sdiag(-MI.diagonal() ** 2) * Av.T * A)
-
-        else:  # isotropic, variable in space
-            if not invert_matrix and not invert_model:
-                dMdprop = n_elements * Av.T * A
-            elif invert_matrix and invert_model:
-                dMdprop = n_elements * (
-                    sdiag(MI.diagonal() ** 2) * Av.T * A * sdiag(1.0 / model**2)
-                )
-            elif invert_model:
-                dMdprop = n_elements * Av.T * A * sdiag(-1.0 / model**2)
-            elif invert_matrix:
-                dMdprop = n_elements * (sdiag(-MI.diagonal() ** 2) * Av.T * A)
-
-        if dMdprop is not None:
-
-            def innerProductDeriv(v=None):
-                if v is None:
-                    warnings.warn(
-                        "Depreciation Warning: TensorMesh.innerProductDeriv."
-                        " You should be supplying a vector. "
-                        "Use: sdiag(u)*dMdprop",
-                        FutureWarning,
-                        stacklevel=2,
-                    )
-                    return dMdprop
-                return sdiag(v) * dMdprop
-
-            return innerProductDeriv
-        else:
-            return None
-
-    def _fastInnerProductLine(
-        self, model=None, invert_model=False, invert_matrix=False
-    ):
-        """Fast version of get_face_inner_product_deriv.
-
-        This does not handle the case of a full tensor property.
-
-        Parameters
-        ----------
-        model : numpy.ndarray
-            material property (tensor properties are possible) at each cell center (nC, (1, 3, or 6))
-        invert_model : bool
-            inverts the material property
-        invert_matrix : bool
-            inverts the matrix
-
-        Returns
-        -------
-        (n_edges, n_edges) scipy.sparse.csr_matrix
-            M, the inner product matrix
-
-        """
-        if model is None:
-            model = np.ones(self.nE)
-
-        if invert_model:
-            model = 1.0 / model
-
-        if is_scalar(model):
-            model = model * np.ones(self.nE)
-
-        # number of elements we are averaging (equals dim for regular
-        # meshes, but for cyl, where we use symmetry, it is 1 for edge
-        # variables and 2 for face variables)
-        # if self._meshType == "CYL":
-        #     shape = getattr(self, "vn" + projection_type)
-        #     n_elements = sum([1 if x != 0 else 0 for x in shape])
-        # else:
-        #     n_elements = self.dim - 1
-
-        # Isotropic case only
-        if model.size == self.nE:
-            Lprop = self.edge_lengths * mkvc(model)
-            M = sdiag(Lprop)
-
-        else:
-            raise Exception(
-                "Unexpected shape of tensor: {}".format(model.shape),
-                "Must be scalar or have length equal to total number of edges.",
-            )
-
-        if invert_matrix:
-            return sdinv(M)
-        else:
-            return M
-
-    def _fastInnerProductLineDeriv(
-        self, model, invert_model=False, invert_matrix=False
-    ):
-        """Faster function for inner product derivatives on tensor meshes.
-
-        Parameters
-        ----------
-        model : numpy.ndarray
-            material property (tensor properties are possible) at each cell center (nC, (1, 3, or 6))
-        invert_model : bool
-            inverts the material property
-        invert_matrix : bool
-            inverts the matrix
-
-        Returns
-        -------
-        function
-            dMdmu, the derivative of the inner product matrix
-        """
-        if is_scalar(model):
-            tensorType = 0
-        elif model.size == self.nE:
-            tensorType = 1
-        else:
-            raise Exception(
-                "Unexpected shape of tensor: {}.".format(model.shape),
-                "Must be scalar or have length equal to total number of edges: {}.".format(
-                    self.nE
-                ),
-            )
-
-        dMdprop = None
-
-        if invert_matrix or invert_model:
-            MI = self._fastEdgePropertiesInnerProduct(
-                model,
-                invert_model=invert_model,
-                invert_matrix=invert_matrix,
-            )
-
-        L = sdiag(self.edge_lengths)
-        if tensorType == 0:  # isotropic, constant
-            ones = sp.csr_matrix(
-                (np.ones(self.nE), (range(self.nE), np.zeros(self.nE))),
-                shape=(self.nE, 1),
-            )
-            if not invert_matrix and not invert_model:
-                dMdprop = L * ones
-            elif invert_matrix and invert_model:
-                dMdprop = sdiag(MI.diagonal() ** 2) * L * ones * sdiag(1.0 / model**2)
-            elif invert_model:
-                dMdprop = L * sdiag(-1.0 / model**2)
-            elif invert_matrix:
-                dMdprop = sdiag(-MI.diagonal() ** 2) * L
-
-        elif tensorType == 1:  # isotropic, variable in space
-            if not invert_matrix and not invert_model:
-                dMdprop = L
-            elif invert_matrix and invert_model:
-                dMdprop = sdiag(MI.diagonal() ** 2) * L * sdiag(1.0 / model**2)
-            elif invert_model:
-                dMdprop = L * sdiag(-1.0 / model**2)
-            elif invert_matrix:
-                dMdprop = sdiag(-MI.diagonal() ** 2) * L
-
-        elif tensorType == 2:  # anisotropic
-            raise NotImplementedError(
-                "EdgePropertiesInnerProductDeriv not implemented for anisotropy."
-            )
-
-        if dMdprop is not None:
-
-            def innerProductDeriv(v=None):
-                if v is None:
-                    warnings.warn(
-                        "Depreciation Warning: TensorMesh.innerProductDeriv."
-                        " You should be supplying a vector. "
-                        "Use: sdiag(u)*dMdprop",
-                        FutureWarning,
-                        stacklevel=2,
-                    )
-                    return dMdprop
-                return sdiag(v) * dMdprop
-
-            return innerProductDeriv
-        else:
-            return None
-
     # DEPRECATED
     @property
     def hx(self):
